@@ -1,14 +1,10 @@
 {
-  "name": "apache-arrow",
   "version": "0.2.0",
-  "types": "./types/Arrow.d.ts",
-  "typings": "./types/Arrow.d.ts",
-  "main": "./targets/es5/cjs/Arrow.js",
-  "module": "./targets/es5/esm/Arrow.js",
-  "browser": "./targets/es5/umd/Arrow.js",
-  "jsnext:main": "./targets/es2015/esm/Arrow.js",
-  "esnext:main": "./targets/esnext/esm/Arrow.js",
+  "name": "apache-arrow",
   "description": "Apache Arrow columnar in-memory format",
+  "bin": {
+    "arrow2csv": "bin/arrow2csv"
+  },
   "scripts": {
     "lerna": "lerna",
     "test": "gulp test",
@@ -57,50 +53,54 @@
     "command-line-usage": "4.0.1"
   },
   "dependencies": {
+    "command-line-args": "4.0.7",
+    "command-line-usage": "4.0.2",
     "flatbuffers": "trxcllnt/flatbuffers-esm",
+    "json-bignum": "0.0.3",
     "text-encoding-utf-8": "^1.0.2",
-    "tslib": "^1.8.0"
+    "tslib": "1.8.1"
   },
   "devDependencies": {
-    "@std/esm": "0.13.0",
+    "@std/esm": "0.18.0",
     "@types/flatbuffers": "1.6.5",
-    "@types/glob": "5.0.33",
-    "@types/jest": "21.1.6",
-    "@types/node": "8.0.51",
+    "@types/glob": "5.0.34",
+    "@types/jest": "21.1.8",
+    "@types/node": "8.5.0",
     "@types/text-encoding": "0.0.32",
+    "ast-types": "0.10.1",
     "benchmark": "2.1.4",
-    "command-line-args": "4.0.7",
     "coveralls": "3.0.0",
     "del": "3.0.0",
     "esdoc": "1.0.3",
     "esdoc-standard-plugin": "1.0.0",
     "glob": "7.1.2",
-    "google-closure-compiler": "20171112.0.0",
+    "google-closure-compiler": "20171203.0.0",
     "gulp": "github:gulpjs/gulp#4.0",
     "gulp-json-transform": "0.4.5",
     "gulp-rename": "1.2.2",
     "gulp-sourcemaps": "2.6.1",
+    "gulp-transform-js-ast": "1.0.2",
     "gulp-typescript": "3.2.3",
-    "ix": "2.3.1",
+    "ix": "2.3.4",
     "jest": "21.2.1",
     "jest-environment-node-debug": "2.0.0",
     "json": "9.0.6",
     "lerna": "2.5.1",
-    "lint-staged": "5.0.0",
+    "lint-staged": "6.0.0",
     "merge2": "1.2.0",
     "mkdirp": "0.5.1",
     "npm-run-all": "4.1.2",
     "pump": "1.0.2",
     "rimraf": "2.6.2",
-    "rxjs": "5.5.2",
+    "rxjs": "5.5.5",
     "shx": "0.2.2",
     "source-map-loader": "0.2.3",
-    "trash": "4.1.0",
-    "ts-jest": "21.2.1",
+    "trash": "4.2.1",
+    "ts-jest": "21.2.4",
     "tslint": "5.8.0",
-    "typescript": "2.6.1",
-    "uglifyjs-webpack-plugin": "1.0.1",
-    "webpack": "3.8.1",
+    "typescript": "2.6.2",
+    "uglifyjs-webpack-plugin": "1.1.2",
+    "webpack": "3.10.0",
     "xml2js": "0.4.19"
   },
   "lint-staged": {
