@@ -868,8 +868,8 @@ class IntegrationRunner(object):
         self.debug = debug
 
     def run(self):
-        for producer, consumer in itertools.product(self.testers,
-                                                    self.testers):
+        for producer, consumer in itertools.product(filter(lambda t: t.PRODUCER, self.testers),
+                                                    filter(lambda t: t.CONSUMER, self.testers)):
             self._compare_implementations(producer, consumer)
 
     def _compare_implementations(self, producer, consumer):
@@ -909,6 +909,8 @@ class IntegrationRunner(object):
 
 
 class Tester(object):
+    PRODUCER = False
+    CONSUMER = False
 
     def __init__(self, debug=False):
         self.debug = debug
@@ -927,6 +929,8 @@ class Tester(object):
 
 
 class JavaTester(Tester):
+    PRODUCER = True
+    CONSUMER = True
 
     _arrow_version = load_version_from_pom()
     ARROW_TOOLS_JAR = os.environ.get(
@@ -978,6 +982,8 @@ class JavaTester(Tester):
 
 
 class CPPTester(Tester):
+    PRODUCER = True
+    CONSUMER = True
 
     EXE_PATH = os.environ.get(
         'ARROW_CPP_EXE_PATH',
@@ -1025,6 +1031,41 @@ class CPPTester(Tester):
             print(cmd)
         os.system(cmd)
 
+class JSTester(Tester):
+    PRODUCER = False
+    CONSUMER = True
+
+    INTEGRATION_EXE = os.path.join(ARROW_HOME, 'js/bin/integration.js')
+
+    name = 'JS'
+
+    def _run(self, arrow_path=None, json_path=None, command='VALIDATE'):
+        cmd = [self.INTEGRATION_EXE]
+
+        if arrow_path is not None:
+            cmd.extend(['-a', arrow_path])
+
+        if json_path is not None:
+            cmd.extend(['-j', json_path])
+
+        cmd.extend(['--mode', command])
+
+        if self.debug:
+            print(' '.join(cmd))
+
+        run_cmd(cmd)
+
+    def validate(self, json_path, arrow_path):
+        return self._run(arrow_path, json_path, 'VALIDATE')
+
+    def stream_to_file(self, stream_path, file_path):
+        # Just copy stream to file, we can read the stream directly
+        cmd = ['cp', stream_path, file_path]
+        cmd = ' '.join(cmd)
+        if self.debug:
+            print(cmd)
+        os.system(cmd)
+
 
 def get_static_json_files():
     glob_pattern = os.path.join(ARROW_HOME, 'integration', 'data', '*.json')
@@ -1032,7 +1073,7 @@ def get_static_json_files():
 
 
 def run_all_tests(debug=False):
-    testers = [CPPTester(debug=debug), JavaTester(debug=debug)]
+    testers = [CPPTester(debug=debug), JavaTester(debug=debug), JSTester(debug=debug)]
     static_json_files = get_static_json_files()
     generated_json_files = get_generated_json_files()
     json_files = static_json_files + generated_json_files
