@@ -27,11 +27,12 @@ const npmOrgName = `@${npmPkgName}`;
 const releasesRootDir = `targets`;
 const knownTargets = [`es5`, `es2015`, `esnext`];
 const knownModules = [`cjs`, `esm`, `cls`, `umd`];
-const moduleFormatsToSkipCombosOf = { cls: true };
+const moduleFormatsToSkipCombosOf = { cls: { test: true, integration: true } };
 const packageJSONFields = [
   `version`, `license`, `description`,
   `author`, `homepage`, `repository`,
-  `bugs`, `keywords`,  `dependencies`
+  `bugs`, `keywords`,  `dependencies`,
+  `bin`
 ];
 
 const metadataFiles = [`LICENSE.txt`, `NOTICE.txt`, `README.md`].map((filename) => {
