@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-import * as Schema_ from '../format/Schema';
+import * as Schema_ from '../format/fb/Schema';
 import Type = Schema_.org.apache.arrow.flatbuf.Type;
 
 export interface Vector<T = any> extends Iterable<T | null> {
@@ -29,6 +29,7 @@ export interface Vector<T = any> extends Iterable<T | null> {
     concat(...vectors: Vector<T>[]): Vector<T>;
     slice<R = T[]>(start?: number, end?: number): R;
 }
+
 export class Vector<T = any> implements Vector<T> {
     slice<R = T[]>(start?: number, end?: number): R {
         let { length } = this, from = start! | 0;
